\name{FHtab}
\docType{data}
\alias{FHtab}
\title{Intergenerational mobility table}
\usage{data(FHtab)}
\description{
The \code{FHtab} data frame has 25 rows and 3 columns.
\code{OccFather} is father's occupation, \code{OccSon} is son's occupation, both have 5 categories:
\itemize{
	\item Upper nonmanual
	\item Lower nonmanual
	\item Upper manual
	\item Lower manual
	\item Farm
}
\code{Freq} contains the table frequencies
}
\source{
 Table from page 11 of \emph{Mobility Tables}. Original source: page 49 of \emph{Opportunity and Change}.
}
\references{
  Hout, Michael. (1983). \cite{Mobility Tables}. Sage Publication 07-03

  Featherman, David L., Robert M. Hauser. (1978). \cite{Opportunity and change}. New York: Academic Press.
}
\keyword{datasets}
