\name{logan}
\docType{data}
\alias{logan}
\title{Data from the 1972-78 GSS data used by Logan (1983: 332-333)}
\usage{data(logan)}
\description{
Intergenerational occupational mobility data with covariates.
}
\format{
  A data frame with 838 observations on the following 4 variables.
  \describe{
    \item{occ}{a factor with levels \code{farm}, \code{operatives}, \code{craftsmen}, \code{sales}, and \code{professional}}
    \item{focc}{a factor with levels \code{farm}, \code{operatives}, \code{craftsmen}, \code{sales}, and \code{professional}}
    \item{educ}{a numeric vector}
    \item{black}{a factor with levels \code{non-black} and \code{black}}
  }
}
\source{
 1972-78 GSS (cf. Logan, 1983: 332-333)
}
\references{
Logan, John A. (1983). A Multivariate Model for Mobility Tables. \cite{American Journal of Sociology} 89: 324-349.}
\keyword{datasets}
