\name{mclgen}

\alias{mclgen}

\title{Deprecated}

\description{
\code{mclgen} has been deprecated. Use \code{\link[mlogit]{mlogit.data}} in the \code{mlogit} package
	
\code{mclgen} was used to restructure a data-frame into a \emph{person-choice}
file for estimation of a multinomial logit as a conditional logit model. The \code{\link[mlogit]{mlogit.data}} does this as well but has greater flexibility. See examples in the \code{\link{sqtab}} help file.
}


