\name{FHtab}
\alias{FHtab}
\non_function{}
\title{Intergenerational mobility table}
\usage{data(FHtab)}
\description{
The \code{FHtab} data frame has 25 rows and 3 columns.
\code{OccFather} is father's occupation, \code{OccSon} is son's occupation, both have 5 categories (\emph{Upper nonmanual}, \emph{Lower nonmanual}, \emph{Upper manual}, \emph{Lower manual}, and \emph{Farm}. \code{Freq} contains the table frequencies
}
\source{
 Table from page 11 of \emph{Mobility Tables}. Original source:
 page 49 of \emph{Opportunity and Change}.
}
\references{
  Hout, Michael. (1983). \emph{Mobility Tables}. Sage Publication 07-03

  Featherman, David L., Robert M. Hauser. (1978).
  \emph{Opportunity and change}. New York: Academic Press.
}
\keyword{datasets}
