\name{FHtab}
\docType{data}
\alias{FHtab}
\title{Intergenerational mobility table}
\description{
A two-way table of father's occupation (\code{OccFather}) by son's occupation (\code{OccSon}) both of which have the following 5 categories:
\itemize{
	\item Upper nonmanual
	\item Lower nonmanual
	\item Upper manual
	\item Lower manual
	\item Farm
}
The table contains 19912 cases.
}
\usage{data(FHtab)}
\format{A \code{table} object}
\source{
Table from page 11 of \emph{Mobility Tables}. Original source: page 49 of \emph{Opportunity and Change}.
}
\references{
Hout, Michael. (1983). \cite{Mobility Tables}. Sage Publication 07-03

Featherman, David L., Robert M. Hauser. (1978). \cite{Opportunity and change}. New York: Academic Press.
}
\keyword{datasets}
